\name{samplezmember}
\alias{samplezmember}
\title{
Update individual level component indezes.
}
\description{
Update individual level component indezes.
}
\usage{
samplezmember(phi, data, w, zHH, serial)
}
\arguments{
  \item{phi}{
  phi parameter.
}
  \item{data}{
  input data.
}
  \item{w}{
  w parameters.
}
  \item{zHH}{
  household level component indexes.
}
  \item{serial}{
  Initial household indexes.
}
}

\value{
A vector for the updated individual level component indexes.
}

\author{
Quanli Wang
}
\keyword{sampler}
\keyword{individual level}
